fix: adjust arch for aws cli

+++ Dockerfile
@@ -123,8 +123,10 @@ RUN ARCH=$(dpkg --print-architecture | sed -e 's/armhf/arm/') && \
     etcdutl version
 
 # aws-cli
-RUN mkdir -p /tmp/aws-cli-download && \
-    curl "https://awscli.amazonaws.com/awscli-exe-linux-$(uname -m).zip" -o "/tmp/aws-cli-download/awscliv2.zip" && \
+RUN ARCH=$(dpkg --print-architecture | sed -e 's/arm64/aarch64/' -e 's/armhf/arm/') && \
+    if [ "$ARCH" = "arm" ] ; then exit 0 ; fi && \
+    mkdir -p /tmp/aws-cli-download && \
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-${ARCH}.zip" -o "/tmp/aws-cli-download/awscliv2.zip" && \
     unzip -q /tmp/aws-cli-download/awscliv2.zip -d /tmp/aws-cli-download && \
     /tmp/aws-cli-download/aws/install && \
     rm -rf /tmp/aws-cli-download && \
